feat: added support for opting out of setting injection.
The function that parses a conan recipe and injects settings from the
configuration during the export step now checks for 'injected_*' version
of the setting first. If injected_name exists for example, then it is
replaced and name is left untouched.

This allows a conan recipe to opt out of having a setting (such as name
or version) overwritten by defining an injected_* version of the
setting.